tesseract-tester/requirements.txt
opencv-python
pillow
pytesseract
matplotlib